Changes Ticket, in advance of JPA there is only one class for tickets now.
To different between the two types the already available field "type" is used. The Database was changed so that there is only one table now. This table was extended by the missing field "estimated_time" whose value will be NULL or an empty string in case of type == bug. The Database operations on tickets should now be faster cause there is no need to compare two tables in order to differenent between an feature or bug ticket.