deleteTicket action fixed
I wasn't able to delete my tickets. After a lot of debugging and
enabling JPA Logging I figured out that my ticket was removed and for
some reason after this inserted again. So in fact the ticket wasn't
deleted at all.
I added a Cascade anotations to the Joined Fields in Ticket to ensure
the transaction didn't stock cause of an inconstistant datasource after
letting a few joins left. This was nice but it didn't solved the
problem.
Actualy while debugging every line and looking at the JPA Queries
executed I found out that the addional and quite unnessessary
removeTCRelation statement was the problem. After removing this
everything was fine - after at least 8 h of debugging :(

The problem was caused by the fact that I reused my already deleted
ticket and merged it DBManager.removeTCRelation to ensure it is in sync
with the database. Unfourtunatly I readded the Ticket this way an
removed all Ticket-Component-Relations - althrou there weren't none
cause they were properly deleted by the cascading delete of
deleteTicket.

What the fuck!!